third_party/WebKit/LayoutTests/fast/html/imports/import-insert.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<script src="resources/import-helpers.js"></script>
<head>
</head>
<body>
<script>

waitAndTest([
    {
        description: "It shouldn't be loaded before insertion.",
        setup: function()
        {
            window.placeholder = createPlaceholder();
            window.toInsert = document.createElement("link");
            window.toInsert.setAttribute("rel", "import");
            window.toInsert.setAttribute("href", "resources/hello.html");

        },
        ready: function() { return window.placeholder.import; },
        test: function()
        {
            shouldBeNull("toInsert.import");
        }
    },
    {
        description: "It should be loaded once it is inserted.",
        setup: function()
        {
            document.head.appendChild(window.toInsert);
        },
        ready: function() { return toInsert.import && toInsert.import.querySelector('h1'); },
        test: function()
        {
            shouldBe("toInsert.import.querySelector('h1').innerHTML", "'Hello'");
        }
    }
]);

</script>
</body>
</html>
